Move proposal template vocabulary.
